factor out material ui a table and user preference
